Split useAutoClose out into a separate file
