+++ src/feedback/FeedbackDialog.tsx
@@ -1,5 +1,5 @@
 import { Fragment, FunctionComponent, h } from "preact";
-import { useCallback, useEffect, useRef, useState } from "preact/hooks";
+import { useCallback, useEffect, useRef } from "preact/hooks";
 
 import { Close } from "./icons/Close";
 import { Logo } from "./icons/Logo";
@@ -19,6 +19,7 @@ import {
   OpenFeedbackFormOptions,
   WithRequired,
 } from "./types";
+import { useAutoClose } from "./hooks/useAutoClose";
 
 type Position = Partial<
   Record<"top" | "left" | "right" | "bottom", number | string>
@@ -41,71 +42,6 @@ const DEFAULT_TRANSLATIONS: FeedbackTranslations = {
   ScoreVerySatisfiedLabel: "Very satisfied",
   SuccessMessage: "Thank you for sending your feedback!",
   SendButton: "Send",
-};
-
-const useAutoClose = ({
-  initialDuration,
-  enabled,
-  onEnd,
-}: {
-  initialDuration: number;
-  enabled: boolean;
-  onEnd: () => void;
-}): {
-  duration: number;
-  elapsedFraction: number;
-  startTime: number;
-  endTime: number;
-  stopped: boolean;
-  startWithDuration: (duration: number) => void;
-  stop: () => void;
-} => {
-  const [stopped, setStopped] = useState(!enabled);
-  const [duration, setDuration] = useState(initialDuration);
-  const [startTime, setStartTime] = useState(Date.now());
-  const [currentTime, setCurrentTime] = useState(Date.now());
-
-  useEffect(() => {
-    if (stopped) return;
-
-    const t = setInterval(() => {
-      setCurrentTime(Date.now());
-
-      if (Date.now() >= startTime + duration) {
-        clearTimeout(t);
-        setStopped(true);
-        onEnd();
-      }
-    }, 100);
-
-    return () => {
-      clearTimeout(t);
-    };
-  }, [stopped]);
-
-  const stop = useCallback(() => {
-    setStopped(true);
-  }, []);
-
-  const startWithDuration = useCallback((nextDuration: number) => {
-    setStartTime(Date.now());
-    setDuration(nextDuration);
-    setStopped(false);
-  }, []);
-
-  const endTime = startTime + duration;
-  const elapsedMs = stopped ? 0 : currentTime - startTime;
-  const elapsedFraction = elapsedMs / duration;
-
-  return {
-    duration,
-    elapsedFraction,
-    startTime,
-    endTime,
-    stopped,
-    startWithDuration,
-    stop,
-  };
 };
 
 const RadialProgress: FunctionComponent<{
